improve separation of concepts
