+++ src/main/scala/org/gammf/collabora/util/Collaboration.scala
@@ -2,6 +2,7 @@ package org.gammf.collabora.util
 
 import org.gammf.collabora.util.CollaborationRight.CollaborationRight
 import org.gammf.collabora.util.CollaborationType.CollaborationType
+import reactivemongo.bson.{BSONDocument, BSONDocumentReader, BSONDocumentWriter}
 
 trait Collaboration {
 
@@ -15,6 +16,30 @@ trait Collaboration {
 }
 
 case class CollaborationUser(user: String, right: CollaborationRight)
+
+object CollaborationUser {
+
+  implicit object BSONtoCollaborationUser extends BSONDocumentReader[CollaborationUser] {
+    def read(doc: BSONDocument): CollaborationUser = {
+      CollaborationUser(
+        user = doc.getAs[String]("user").get,
+        right = doc.getAs[String]("right").map(r => CollaborationRight.withName(r)).get
+      )
+    }
+  }
+
+
+  implicit object CollaborationUserToBSON extends BSONDocumentWriter[CollaborationUser] {
+    def write(user: CollaborationUser): BSONDocument = {
+      BSONDocument(
+        "user" -> user.user,
+        "right" -> user.right.toString
+      )
+    }
+  }
+
+
+}
 
 object CollaborationType extends Enumeration {
   type CollaborationType = Value
